test(e2e): accept Codex capabilities before onboarding
Use the shared fixture consent flow before non-interactive onboarding in
Codex's complete package lane. Keep real doctor CLI checks in the Docker
scenario so metadata-only package fixtures remain independently runnable.
